test(web): split trace api server tests
